requirements.txt
cryptography==37.0.2
Flask==1.1.2
pandas==1.4.2
psycopg2==2.9.3
psycopg2_binary==2.9.1
python-dotenv==0.20.0
SQLAlchemy==1.4.36
